+++ CHANGELOG.md
@@ -22,7 +22,7 @@
     - [Todo] Use GA version of Keypin `0.7.3`
     - For parser function `keypin.util/any->fn`
 - Key definitions
-  - [TODO - BREAKING CHANGE] Switch launcher from config to context
+  - [BREAKING CHANGE] Switch launcher key from config `"bract.core.launcher"` to context `:bract.core/launcher`
   - Change launcher parser from `str->var` to `any->fn`
 - [Todo - BREAKING CHANGE] Drop bract.core.type and bract.core.impl in favour of a multimethod (for flexibility)
 - [Todo - BREAKING CHANGE] Switch launcher from config to context value